dashboard/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
altair>=4.0.0
azure-storage-blob>=12.0.0
azure-core>=1.20.0
python-dotenv>=0.19.0
streamlit-option-menu>=0.3.0
streamlit-aggrid>=0.3.0
streamlit-plotly-events>=0.0.6
streamlit-extras>=0.3.0
seaborn>=0.11.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
